requirements.txt
websockets>=12.0
aiohttp>=3.9.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
python-dotenv>=1.0.0

# Telegram modülleri (monitoring/)
apscheduler>=3.10.0
python-telegram-bot==20.7

# Monetizasyon (opsiyonel — bot_manager, subscription, webhook)
# supabase>=2.0.0
# fastapi>=0.100.0
# uvicorn>=0.23.0

# AI analiz (opsiyonel — ai_analyst)
# ollama>=0.1.0
